Remove @glimmer/component dependency
Previously we had a circular dependency between @glimmer/component and @glimmer/application, at least when it came to tests. This made it somewhat difficult to coordinate releases.

This commit removes the @glimmer/component dependency from this package. Instead, we inline a stripped down implementation of ComponentManager/ComponentDefinition—just enough to let us test rendering end to end. We can keep more exhaustive rendering tests in @glimmer/component, which can now easily depend on this package as a development dependency for testing.