{{action}} shouldn’t error in template-only components
This was trying to get the component to bind the action function, but template-only components may not have a component object and be passing through a function that has already been bound. Now we guard for the existence of the componentRef before trying to get its value. We also now return null for template-only components from `getSelf()`.